backend/requirements.txt
# Django Backend Dependencies
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
pandas>=2.0
reportlab>=4.0
openpyxl>=3.1
